tidy(ui): tidy connection validation functions and logic
