--- invokeai/frontend/web/src/features/nodes/store/util/validateSourceAndTargetTypes.ts
@@ -1,90 +1,0 @@
-import { type FieldType, isStatefulFieldType } from 'features/nodes/types/field';
-import { isEqual, omit } from 'lodash-es';
-
-export const areTypesEqual = (sourceType: FieldType, targetType: FieldType) => {
-  const _sourceType = isStatefulFieldType(sourceType) ? omit(sourceType, 'originalType') : sourceType;
-  const _targetType = isStatefulFieldType(targetType) ? omit(targetType, 'originalType') : targetType;
-  const _sourceTypeOriginal = isStatefulFieldType(sourceType) ? sourceType.originalType : sourceType;
-  const _targetTypeOriginal = isStatefulFieldType(targetType) ? targetType.originalType : targetType;
-  if (isEqual(_sourceType, _targetType)) {
-    return true;
-  }
-  if (_targetTypeOriginal && isEqual(_sourceType, _targetTypeOriginal)) {
-    return true;
-  }
-  if (_sourceTypeOriginal && isEqual(_sourceTypeOriginal, _targetType)) {
-    return true;
-  }
-  if (_sourceTypeOriginal && _targetTypeOriginal && isEqual(_sourceTypeOriginal, _targetTypeOriginal)) {
-    return true;
-  }
-  return false;
-};
-
-/**
- * Validates that the source and target types are compatible for a connection.
- * @param sourceType The type of the source field.
- * @param targetType The type of the target field.
- * @returns True if the connection is valid, false otherwise.
- */
-export const validateSourceAndTargetTypes = (sourceType: FieldType, targetType: FieldType) => {
-  // TODO: There's a bug with Collect -> Iterate nodes:
-  // https://github.com/invoke-ai/InvokeAI/issues/3956
-  // Once this is resolved, we can remove this check.
-  if (sourceType.name === 'CollectionField' && targetType.name === 'CollectionField') {
-    return false;
-  }
-
-  if (areTypesEqual(sourceType, targetType)) {
-    return true;
-  }
-
-  /**
-   * Connection types must be the same for a connection, with exceptions:
-   * - CollectionItem can connect to any non-Collection
-   * - Non-Collections can connect to CollectionItem
-   * - Anything (non-Collections, Collections, CollectionOrScalar) can connect to CollectionOrScalar of the same base type
-   * - Generic Collection can connect to any other Collection or CollectionOrScalar
-   * - Any Collection can connect to a Generic Collection
-   */
-
-  const isCollectionItemToNonCollection = sourceType.name === 'CollectionItemField' && !targetType.isCollection;
-
-  const isNonCollectionToCollectionItem =
-    targetType.name === 'CollectionItemField' && !sourceType.isCollection && !sourceType.isCollectionOrScalar;
-
-  const isAnythingToCollectionOrScalarOfSameBaseType =
-    targetType.isCollectionOrScalar && sourceType.name === targetType.name;
-
-  const isGenericCollectionToAnyCollectionOrCollectionOrScalar =
-    sourceType.name === 'CollectionField' && (targetType.isCollection || targetType.isCollectionOrScalar);
-
-  const isCollectionToGenericCollection = targetType.name === 'CollectionField' && sourceType.isCollection;
-
-  const areBothTypesSingle =
-    !sourceType.isCollection &&
-    !sourceType.isCollectionOrScalar &&
-    !targetType.isCollection &&
-    !targetType.isCollectionOrScalar;
-
-  const isIntToFloat = areBothTypesSingle && sourceType.name === 'IntegerField' && targetType.name === 'FloatField';
-
-  const isIntOrFloatToString =
-    areBothTypesSingle &&
-    (sourceType.name === 'IntegerField' || sourceType.name === 'FloatField') &&
-    targetType.name === 'StringField';
-
-  const isTargetAnyType = targetType.name === 'AnyField';
-
-  // One of these must be true for the connection to be valid
-  return (
-    isCollectionItemToNonCollection ||
-    isNonCollectionToCollectionItem ||
-    isAnythingToCollectionOrScalarOfSameBaseType ||
-    isGenericCollectionToAnyCollectionOrCollectionOrScalar ||
-    isCollectionToGenericCollection ||
-    isIntToFloat ||
-    isIntOrFloatToString ||
-    isTargetAnyType
-  );
-};
